requirements.txt
# requirements.txt

aiosmtplib>=4.0.0
asgiref>=3.8.1
astroid>=3.3.9
attrs>=25.3.0
black>=25.1.0
blinker>=1.9.0
Brotli>=1.1.0
cachelib>=0.13.0
certifi>=2025.1.31
charset-normalizer>=3.4.1
click>=8.1.8
colorama>=0.4.6
coverage>=7.7.0
cssbeautifier>=1.15.4
dill>=0.3.9
djlint>=1.36.4
dnspython>=2.7.0
EditorConfig>=0.17.0
et_xmlfile>=2.0.0
exceptiongroup>=1.2.2
execnet>=2.1.1
Flask>=3.1.0
Flask-Caching>=2.3.1
Flask-Compress>=1.17
gunicorn>=23.0.0
h11>=0.14.0
html-tag-names>=0.1.2
html-void-elements>=0.1.0
idna>=3.10
iniconfig>=2.0.0
isort>=6.0.1
itsdangerous>=2.2.0
Jinja2>=3.1.6
jsbeautifier>=1.15.4
json5>=0.10.0
MarkupSafe>=3.0.2
mccabe>=0.7.0
mypy-extensions>=1.0.0
numpy>=2.2.4
openpyxl>=3.1.5
outcome>=1.3.0.post0
packaging>=24.2
pandas>=2.2.3
pandas-stubs>=2.2.3.250308
passlib>=1.7.4
pathspec>=0.12.1
platformdirs>=4.3.6
pluggy>=1.5.0
pylint>=3.3.5
pymongo>=4.11.2
PySocks>=1.7.1
pytest>=8.3.5
pytest-cov>=6.0.0
pytest-mock>=3.14.0
pytest-xdist>=3.6.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2025.1
PyYAML>=6.0.2
regex>=2024.11.6
requests>=2.32.3
selenium>=4.29.0
six>=1.17.0
sniffio>=1.3.1
sortedcontainers>=2.4.0
tomli>=2.2.1
tomlkit>=0.13.2
tqdm>=4.67.1
trio>=0.29.0
trio-websocket>=0.12.2
types-colorama>=0.4.15.20240311
types-passlib>=1.7.7.20241221
types-pytz>=2025.1.0.20250204
typing_extensions>=4.12.2
tzdata>=2025.1
urllib3>=2.3.0
webdriver-manager>=4.0.2
websocket-client>=1.8.0
Werkzeug>=3.1.3
wsproto>=1.2.0
zstandard>=0.23.0
